requirements.txt
transformers
faiss-gpu
faiss-cpu
sentence-transformers
datasets
sentencepiece
transformers[sentencepiece]